requirements.txt
numpy==1.26.4
pandas==2.1.4
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
msgpack==1.0.7
numba==0.58.1
uvloop==0.19.0
ta==0.11.0
pandas-ta==0.3.14b0
scipy==1.11.4
torch==2.1.2
scikit-learn==1.3.2
xgboost==2.0.2
stable-baselines3==2.1.0
deap==1.4.1
bayesian-optimization==1.4.3
vectorbt==0.24.3
backtrader==1.9.78.123
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
aiohttp==3.9.1
pyzmq==25.1.2
websockets==12.0
MetaTrader5==5.0.45
llama-cpp-python==0.2.11
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0
faiss-cpu==1.7.4
chromadb==0.4.22
sqlalchemy==2.0.23
tenacity==8.2.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
mypy==1.8.0
ruff==0.1.7
pre-commit==3.5.0
types-PyYAML==6.0.12.12
pytz==2023.3.post1
